requirements.txt
python-ldap==2.4.12
Flask==0.10.1
Flask-Login==0.2.7
Flask-SQLAlchemy==0.16